unfinished ver 0.8.20.1

+++ demo-workspace/testDemo/src/demo/QuestionTest.java
@@ -4,7 +4,7 @@ import java.util.*;
 
 public class QuestionTest {
 	public static void main(String[] args) {
-		System.out.println(Long.MAX_VALUE);
+		System.out.println(Float.MAX_VALUE);
 		scanTest();
 	}
